Do the rest of the splitting
