+++ src/test_c_kzg_4844.c
@@ -2,6 +2,9 @@
  * This file contains unit tests for C-KZG-4844.
  */
 #include "c_kzg_4844.c"
+#include "c_kzg_7594.c"
+#include "common.c"
+
 #include "tinytest.h"
 
 #include <assert.h>
